backend/requirements.txt
yfinance>=0.2.50
pandas==2.2.2
ta==0.11.0
SQLAlchemy==2.0.30
fastapi==0.111.0
uvicorn==0.30.1
apscheduler==3.10.4
python-dotenv==1.0.1
numpy==1.26.4
requests==2.32.3
